standardized style. moved weather, plants, and animals outside the Auth0 so users can see those things without signing in. updated footer to appear on all pages.
