+++ front_end/src/App.js
@@ -4,15 +4,15 @@ import { useAuth0 } from '@auth0/auth0-react';
 
 import Header from './components/nav/Header.js';
 import Home from './components/pages/Home.js'
-import Search from './components/pages/search/Search.js';
 import Animals from './components/pages/animals/Animals.js';
 import Plants from './components/pages/plants/Plants.js';
 import Weather from './components/pages/Weather.js';
-import About from './components/pages/About.js';
+import About from './components/pages/about/About.js';
 import Trip from './components/pages/Trip.js';
 import LoginButton from './components/auth/LoginButton.js';
 import LogoutButton from './components/auth/LogoutButton.js';
 import Profile from './components/auth/Profile.js';
+import Footer from './components/nav/Footer.js';
 import './components/nav/Header.css';
 import './components/auth/Auth.css';
 
@@ -25,20 +25,20 @@ function App() {
       <Header/>
       <Routes>
         <Route exact path='/' element={<Home />} />
-        <Route path='/search' element={<Search/>} />
         <Route path='/about' element={<About/>} />
+        <Route path='/weather' element={<Weather/>} />
+        <Route path='/animals' element={<Animals/>} /> 
+        <Route path='/plants' element={<Plants/>} />
         
         {/* Show these pages only if the user is authenticated */}
         {isAuthenticated && (
           <>
             <Route path='/trip' element={<Trip/>} />
-            <Route path='/weather' element={<Weather/>} />
-            <Route path='/animals' element={<Animals/>} /> 
-            <Route path='/plants' element={<Plants/>} />
             <Route path='/profile' element={<Profile/>} />
           </>
         )}
       </Routes>
+        <Footer />
     </Router>
     {error && <p>Authentication Error</p>}
     {!error && isLoading && <p>Loading...</p>}
