+++ front_end/src/components/auth/Auth.css
@@ -8,7 +8,7 @@
     display: flex;
     position: absolute;
     top: 20px;
-    left: 1700px;
+    right: 30px;
   }
 
   .login-button:hover {
