+++ front_end/src/components/nav/Footer.css
@@ -1,0 +1,7 @@
+footer {
+    background-color: #333;
+    color: white;
+    text-align: center;
+    padding: 1em 0;
+    width: 100%;
+  }
