+++ front_end/src/components/nav/Footer.js
@@ -1,0 +1,10 @@
+import React from "react";
+import './Footer.css';
+
+export default function Footer() {
+    return (
+        <footer>
+            <p>&copy; 2024 Trailblazers. All rights reserved.</p>
+        </footer>
+    )
+}
