+++ front_end/src/components/pages/about/About.css
@@ -1,0 +1,5 @@
+section {
+    max-width: 800px;
+    margin: 40px auto;
+    padding: 20px;
+  }
