+++ front_end/src/components/pages/trip/historystyles.css
@@ -1,9 +1,4 @@
 /* historystyles.css */
-
-body {
-    font-family: 'Arial', sans-serif;
-    margin: 20px;
-  }
   
   .parent-container {
     margin-bottom: 20px;
